data/requirements.txt
pandas
sqlalchemy
pyarrow
psycopg2-binary
pymongo
apache-airflow
prefect
dbt-core
great-expectations